refactor: extract mobile host picker into MobileHostFilterActionViewItem
Move the bottom sheet UI from an inline phone-layout branch in the
desktop HostFilterActionViewItem into a proper mobile subclass in
browser/parts/mobile/, following the established mobile architecture:

- Create MobileHostFilterActionViewItem extending the desktop class,
  overriding _showMenu() to show a bottom sheet instead of a context menu
- Remove IsPhoneLayoutContext branching from the desktop component
- Move bottom sheet CSS from hostFilter.css to mobileChatShell.css
- Update hostFilter.contribution.ts to instantiate the mobile variant
  for the MobileTitleBarCenter menu registration
- Add Mobile Component Architecture rules to sessions.instructions.md
