+++ src/vs/sessions/browser/parts/mobile/mobileChatShell.css
@@ -629,3 +629,147 @@
 	min-height: 44px;
 	touch-action: manipulation;
 }
+
+/* ---- Mobile Bottom Sheet Host Picker ---- */
+
+.host-picker-sheet-backdrop {
+	position: fixed;
+	inset: 0;
+	z-index: 10000;
+	background: rgba(0, 0, 0, 0.4);
+	animation: host-picker-backdrop-in 200ms ease-out;
+}
+
+.host-picker-sheet-backdrop.dismissing {
+	animation: host-picker-backdrop-out 200ms ease-in forwards;
+}
+
+@keyframes host-picker-backdrop-in {
+	from { opacity: 0; }
+	to { opacity: 1; }
+}
+
+@keyframes host-picker-backdrop-out {
+	from { opacity: 1; }
+	to { opacity: 0; }
+}
+
+.host-picker-sheet {
+	position: absolute;
+	bottom: 0;
+	left: 0;
+	right: 0;
+	background: var(--vscode-editorWidget-background, var(--vscode-editor-background));
+	border-radius: 16px 16px 0 0;
+	padding: 8px 0 calc(16px + env(safe-area-inset-bottom)) 0;
+	animation: host-picker-sheet-in 250ms ease-out;
+	max-height: 70vh;
+	overflow-y: auto;
+	-webkit-overflow-scrolling: touch;
+}
+
+.host-picker-sheet.dismissing {
+	animation: host-picker-sheet-out 200ms ease-in forwards;
+}
+
+@keyframes host-picker-sheet-in {
+	from { transform: translateY(100%); }
+	to { transform: translateY(0); }
+}
+
+@keyframes host-picker-sheet-out {
+	from { transform: translateY(0); }
+	to { transform: translateY(100%); }
+}
+
+.host-picker-sheet-handle {
+	width: 36px;
+	height: 5px;
+	background: var(--vscode-foreground);
+	opacity: 0.3;
+	border-radius: 3px;
+	margin: 4px auto 12px auto;
+}
+
+.host-picker-sheet-title {
+	font-size: 16px;
+	font-weight: 600;
+	color: var(--vscode-foreground);
+	padding: 0 20px 12px 20px;
+}
+
+.host-picker-sheet-list {
+	display: flex;
+	flex-direction: column;
+}
+
+.host-picker-sheet-item {
+	display: flex;
+	align-items: center;
+	gap: 12px;
+	min-height: 52px;
+	padding: 10px 20px;
+	border: none;
+	background: none;
+	color: var(--vscode-foreground);
+	font-size: 16px;
+	font-family: inherit;
+	cursor: pointer;
+	touch-action: manipulation;
+	text-align: left;
+	width: 100%;
+}
+
+.host-picker-sheet-item:active {
+	background: var(--vscode-list-hoverBackground);
+}
+
+.host-picker-sheet-item:focus-visible {
+	outline: 2px solid var(--vscode-focusBorder);
+	outline-offset: -2px;
+}
+
+.host-picker-sheet-item.selected {
+	color: var(--vscode-textLink-foreground);
+}
+
+.host-picker-sheet-item-icon {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	width: 36px;
+	height: 36px;
+	border-radius: 50%;
+	background: var(--vscode-toolbar-hoverBackground);
+	flex-shrink: 0;
+}
+
+.host-picker-sheet-item-icon .codicon {
+	font-size: 18px;
+}
+
+.host-picker-sheet-item-label {
+	flex: 1;
+	min-width: 0;
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+}
+
+.host-picker-sheet-item-status {
+	font-size: 13px;
+	color: var(--vscode-descriptionForeground);
+	flex-shrink: 0;
+}
+
+.host-picker-sheet-item-check {
+	display: flex;
+	align-items: center;
+	flex-shrink: 0;
+	color: var(--vscode-textLink-foreground);
+}
+
+.host-picker-sheet-item-check .codicon {
+	font-size: 16px;
+}
+}
